requirements.txt
mcp[cli]>=1.6.0
rank-bm25>=0.2.2
python-frontmatter>=1.1.0
